chore(odyssey-react): refactor Button types and fs
