+++ packages/odyssey-react/src/components/Button/Button.test.tsx
@@ -12,7 +12,7 @@
 
 import React from "react";
 import { render, fireEvent } from "@testing-library/react";
-import Button from "./Button";
+import Button from ".";
 
 const button = "button";
 const buttonLabel = "Button Label";
